requirements-nuitka.txt
# VoiceInput System - Nuitka专用依赖
# 包含构建可执行文件所需的所有包

# 核心语音识别
funasr==1.0.22
SpeechRecognition==3.10.0
pyaudio==0.2.11

# 机器学习和音频处理
torch==2.3.1+cpu
torchaudio==2.3.1+cpu
numpy>=1.21.0
onnxruntime>=1.12.0

# 中文数字转换
cn2an>=0.5.17

# GUI框架
tkinter-tooltip==2.1.0

# 数据处理
openpyxl==3.1.2
PyYAML==6.0.1
pandas==2.0.3

# 配置管理
python-dotenv==1.0.0

# 日志和调试
colorlog==6.7.0

# 构建工具 (Nuitka)
nuitka==1.9.2
ordered-set==4.1.0
zstandard==0.21.0

# 其他依赖
psutil==5.9.5
requests==2.31.0
